+++ modules/tools/eval/config.el
@@ -8,7 +8,7 @@
 
 (use-package execrun
   :init
-  (setq execrun-kill-buffer-function #'jcs-maybe-kill-this-buffer)
+  (setq execrun-kill-buffer-function #'jcs-maybe-kill-current-buffer)
 
   (message-clean-mode-add-echo-commands '( execrun-compile)))